requirements.txt
asgiref==3.8.1
Django==4.2.16
django-crum==0.7.9
django-extensions==3.2.3
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
django-simple-history==3.7.0
django-tinymce==4.1.0
pillow==11.0.0
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
